+++ packages/react/src/components/accordion/index.tsx
@@ -1,3 +1,3 @@
 export { Accordion } from './accordion';
-export { SingleOpenAccordionGroup } from './accordion-single-open';
-export { MultiOpenAccordionGroup } from './accordion-multi-open';
+export { SingleOpenAccordionGroup } from './accordion-single';
+export { MultiOpenAccordionGroup } from './accordion-multi';
